Move the clusterregistry server implementation to /pkg.
