+++ cmd/clusterregistry/clusterregistry.go
@@ -26,23 +26,23 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/apiserver/pkg/util/logs"
-	"k8s.io/cluster-registry/cmd/clusterregistry/app"
-	"k8s.io/cluster-registry/cmd/clusterregistry/app/options"
+	"k8s.io/cluster-registry/pkg/clusterregistry"
+	"k8s.io/cluster-registry/pkg/clusterregistry/options"
 
 	"github.com/spf13/pflag"
 )
 
 func main() {
+	logs.InitLogs()
+	defer logs.FlushLogs()
+
 	rand.Seed(time.Now().UTC().UnixNano())
 
 	s := options.NewServerRunOptions()
 	s.AddFlags(pflag.CommandLine)
 	flag.InitFlags()
 
-	logs.InitLogs()
-	defer logs.FlushLogs()
-
-	if err := app.Run(s, wait.NeverStop); err != nil {
+	if err := clusterregistry.Run(s, wait.NeverStop); err != nil {
 		fmt.Fprintf(os.Stderr, "%v\n", err)
 		os.Exit(1)
 	}
