+++ pkg/clusterregistry/authenticator/config.go
