+++ pkg/clusterregistry/options/validation.go
